+++ src/cryptography/verify_signature.py
@@ -27,9 +27,22 @@ def recover_vs(vs: str) -> tuple[VRS, VRS]:
     """
     # cut 0x
     _vs = int.from_bytes(bytearray.fromhex(vs[2:]))
-    s = _vs & 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff
+    s = _vs & 0x7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF
     v = (_vs >> 255) + V_OFFSET
     return v, s
+
+
+def compute_vs(v: int, s: str) -> str:
+    """Returns aggregated _vs value."""
+    if v < V_OFFSET and v not in [0, 1]:
+        raise ValueError('Signature invalid v byte.')
+    if v < V_OFFSET:
+        v += V_OFFSET
+    _vs = bytearray.fromhex(s[2:])
+    if not v % 2:
+        _vs[0] |= 0x80
+
+    return '0x' + _vs.hex()
 
 
 def verify_message_with_signature(data: List[Any], abi: List[str], address: str, vrs: Tuple[VRS, VRS, VRS]) -> bool:
