+++ tests/bots/test_pauser.py
@@ -4,10 +4,10 @@ from unittest.mock import Mock
 import pytest
 import variables
 from bots.pauser import PauserBot
+from cryptography.verify_signature import compute_vs
 
 from tests.conftest import DSM_OWNER
 from tests.fixtures import upgrade_staking_router_to_v2
-from tests.utils.signature import compute_vs
 
 # WARNING: These accounts, and their private keys, are publicly known.
 COUNCIL_ADDRESS = '0x70997970C51812dc3A010C7d01b50e0d17dc79C8'
@@ -75,7 +75,7 @@ def get_pause_message(web3, module_id):
             'r': '0x' + signed.r.to_bytes(32, 'big').hex(),
             's': '0x' + signed.s.to_bytes(32, 'big').hex(),
             'v': signed.v,
-            '_vs': compute_vs(signed.v, '0x' + signed.s.to_bytes(32, 'big').hex())
+            '_vs': compute_vs(signed.v, '0x' + signed.s.to_bytes(32, 'big').hex()),
         },
         'type': 'pause',
     }
@@ -99,7 +99,7 @@ def get_pause_message_v2(web3):
             'r': '0x' + signed.r.to_bytes(32, 'big').hex(),
             's': '0x' + signed.s.to_bytes(32, 'big').hex(),
             'v': signed.v,
-            '_vs': compute_vs(signed.v, '0x' + signed.s.to_bytes(32, 'big').hex())
+            '_vs': compute_vs(signed.v, '0x' + signed.s.to_bytes(32, 'big').hex()),
         },
         'type': 'pause',
     }
@@ -202,10 +202,10 @@ def test_pauser_bot(web3_lido_integration, web3_provider_integration, add_accoun
         msg
         for msg in caplog.messages
         if (
-               "Build `pauseDeposits(19628129, ('0xe37ddb5eadce3a4f03274a102c45a21865f9d6fa03f61fdb98466ae0fd677331', "
-               "'0xb1dff99d6b8ad3402c7237341816568bf2831ddbe0bddd111eebd12a59efcb8c'))` tx."
-           )
-           in msg
+            "Build `pauseDeposits(19628129, ('0xe37ddb5eadce3a4f03274a102c45a21865f9d6fa03f61fdb98466ae0fd677331', "
+            "'0xb1dff99d6b8ad3402c7237341816568bf2831ddbe0bddd111eebd12a59efcb8c'))` tx."
+        )
+        in msg
     ]
 
     pb.execute(latest)
