+++ tests/bots/test_unvetter.py
@@ -3,12 +3,12 @@ from unittest.mock import Mock
 
 import pytest
 from bots.unvetter import UnvetterBot
+from cryptography.verify_signature import compute_vs
 from transport.msg_types.common import get_messages_sign_filter
 from transport.msg_types.unvet import UnvetMessage
 from utils.bytes import from_hex_string_to_bytes
 
 from tests.fixtures import upgrade_staking_router_to_v2
-from tests.utils.signature import compute_vs
 
 # WARNING: These accounts, and their private keys, are publicly known.
 COUNCIL_ADDRESS = '0x70997970C51812dc3A010C7d01b50e0d17dc79C8'
@@ -48,7 +48,7 @@ def get_unvet_message(web3) -> UnvetMessage:
             'r': '0x' + signed.r.to_bytes(32, 'big').hex(),
             's': '0x' + signed.s.to_bytes(32, 'big').hex(),
             'v': signed.v,
-            '_vs': compute_vs(signed.v, '0x' + signed.s.to_bytes(32, 'big').hex())
+            '_vs': compute_vs(signed.v, '0x' + signed.s.to_bytes(32, 'big').hex()),
         },
         'type': 'unvet',
     }
